Typescript Capstone Challenge
TypeScript Essentials Capstone Challenge
Code sample consist of implementation of Observer, Observable and usage of this classes (handling requests). The assignment implies the coverage of the types of the entire code not only Observable & Observer but also request, users objects and observer handlers.

Acceptance criteria
The code sample should be fully covered with types.